pyproject.toml
[tool.poetry]
name = "ad_metrics"
version = "0.1.0"
description = ""
authors = ["Oluwafemi Adenuga <phemmylintry@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.5"
psycopg2-binary = "^2.9.5"
python-decouple = "^3.6"
google-cloud-bigquery = "^3.4.1"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
mypy = "^0.991"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
